move example reclustering algs into example dir
